pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-trading-bot"
version = "5.5.1"
description = "Adaptive Evolution Trading & Hedging Execution Robot"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Gunasekar87"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "torch==2.0.1",
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "onnxruntime==1.16.3",
    "stable-baselines3==2.1.0",
    "shimmy==1.3.0",
    "gymnasium==0.29.1",
    "pyzmq==25.1.1",
    "pandas==2.1.3",
    "numpy<2.0.0",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "supabase==2.8.1",
    "psycopg2-binary==2.9.9",
    "ccxt==4.1.63",
    "MetaTrader5==5.0.45",
    "yfinance==0.2.18",
    "aiosqlite==0.19.0",
    "asyncpg==0.29.0",
    "psutil"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black",
    "isort",
    "flake8"
]

[project.scripts]
aether-bot = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
